climate-policy-radar/requirements.txt
GitPython
beautifulsoup4